kali_extractor/requirements.txt
progressbar2==3.39.2
xmljson==0.2.0
pymongo==3.7.2
requests==2.21.0
